refactor: relocate workspace.ts → src/agent/; absorb assistant-output-directives
- src/legacy/core/workspace.ts → src/agent/workspace.ts (+ test)
- src/legacy/core/assistant-output-directives.ts deleted; the single
  exported string inlined as a module constant in workspace.ts (only
  caller was workspace.ts)
- 2 importers updated: agent-manager.ts, agent-init.ts

Part of #632 segment 3.
